repo clean up, rename, and slight refactor for consistency.
