--- src/components/layout/navigation.tsx
@@ -1,9 +1,0 @@
-import { PlusCircle } from "lucide-react"
-
-export const navigationItems = [
-  {
-    name: "Mint Option",
-    href: "/mint-option",
-    icon: PlusCircle
-  }
-]
