+++ src/lib/clusters/defaultCluster.ts
@@ -14,7 +14,10 @@ export const defaultClusters: Cluster[] = [
       endpoint: clusterApiUrl('devnet'),
       network: ClusterNetwork.Devnet,
     },
-    { name: 'local', endpoint: 'http://localhost:8899' },
+    { 
+      name: 'local', 
+      endpoint: 'http://localhost:8899'
+    },
     {
       name: 'testnet',
       endpoint: clusterApiUrl('testnet'),
