+++ src/lib/option-pricing-model/black-scholes-model.ts
